+++ lib/Echidna/UI.hs
@@ -31,14 +31,15 @@ import UnliftIO.Timeout (timeout)
 import Echidna.Campaign (campaign)
 import Echidna.ABI
 import qualified Echidna.Output.JSON
-import Echidna.Solidity
-import Echidna.Test
+import Echidna.Types.Solidity (SolConf(..))
 import Echidna.Types.Campaign
-import Echidna.Types.Test (SolTest)
+import Echidna.Types.Test (TestConf(..), EchidnaTest)
 import Echidna.Types.Tx (Tx, TxConf)
 import Echidna.Types.World (World)
 import Echidna.UI.Report
 import Echidna.UI.Widgets
+
+import EVM.Dapp (DappInfo)
 
 data UIConf = UIConf { _maxTime       :: Maybe Int
                      , _operationMode :: OperationMode
@@ -84,10 +85,10 @@ isTerminal = liftIO $ (&&) <$> queryTerminal (Fd 0) <*> queryTerminal (Fd 1)
 -- | Set up and run an Echidna 'Campaign' and display interactive UI or
 -- print non-interactive output in desired format at the end
 ui :: ( MonadCatch m, MonadRandom m, MonadReader x m, MonadUnliftIO m
-      , Has SolConf x, Has TestConf x, Has TxConf x, Has CampaignConf x, Has Names x, Has TxConf x, Has UIConf x)
+      , Has SolConf x, Has TestConf x, Has TxConf x, Has CampaignConf x, Has Names x, Has TxConf x, Has UIConf x, Has DappInfo x)
    => VM             -- ^ Initial VM state
    -> World          -- ^ Initial world state
-   -> [SolTest]      -- ^ Tests to evaluate
+   -> [EchidnaTest]      -- ^ Tests to evaluate
    -> Maybe GenDict
    -> [[Tx]]
    -> m Campaign